fixed up add ,delete, select, and update functions . Then need more touching up with the ids. Database seem to working perfectly now. Gradle also no issues after adding the dependencies for sqlite. Add numerous comments still ned to add more. Need to add command line arguments for the employees. Maager should have acces to iventory update add and delete. Employee should be able to selcet and add to meal. COuld make certain meals into a list of items that make up that meal. Then have that add to a bill.

+++ src/main/java/Item.java
@@ -1,19 +1,20 @@
+import java.math.BigDecimal;
 
 public class Item {
 
     //attributes of item
     private String item_name;
     private int item_Id;
-    private int price_per_unit;
+    private BigDecimal price_per_unit;
     private String item_type;
     private int quantity;
 
     //GETTERS AND SETTERS
-    public int getPricePerUnit() {
+    public BigDecimal getPricePerUnit() {
         return price_per_unit;
     }
 
-    public void setPricePerUnit(int price_per_unit) {
+    public void setPricePerUnit(BigDecimal price_per_unit) {
         this.price_per_unit = price_per_unit;
     }
